@@ -12,58 +12,58 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@apollo/client": "^3.4.5",
-    "@babel/core": "^7.15.0",
-    "@babel/eslint-parser": "^7.15.0",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-transform-runtime": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/register": "^7.14.5",
-    "@babel/runtime": "^7.14.8",
-    "autoprefixer": "^10.3.1",
-    "axios": "^0.21.1",
+    "@apollo/client": "^3.5.8",
+    "@babel/core": "^7.17.2",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/plugin-proposal-class-properties": "^7.16.7",
+    "@babel/plugin-transform-runtime": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/register": "^7.17.0",
+    "@babel/runtime": "^7.17.2",
+    "autoprefixer": "^10.4.2",
+    "axios": "^0.25.0",
     "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.2.2",
-    "body-parser": "^1.19.0",
-    "compression-webpack-plugin": "^8.0.1",
-    "connect-mongo": "^4.4.1",
-    "copy-webpack-plugin": "^9.0.1",
-    "css-loader": "^6.2.0",
-    "css-minimizer-webpack-plugin": "^3.0.2",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "express": "^4.17.1",
+    "babel-loader": "^8.2.3",
+    "body-parser": "^1.19.1",
+    "compression-webpack-plugin": "^9.2.0",
+    "connect-mongo": "^4.6.0",
+    "copy-webpack-plugin": "^10.2.4",
+    "css-loader": "^6.6.0",
+    "css-minimizer-webpack-plugin": "^3.4.1",
+    "eslint": "^8.8.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "express": "^4.17.2",
     "express-graphql": "^0.12.0",
     "express-session": "^1.17.2",
-    "graphql": "^15.5.1",
-    "graphql-tag": "^2.12.5",
-    "html-webpack-plugin": "^5.3.2",
+    "graphql": "^16.3.0",
+    "graphql-tag": "^2.12.6",
+    "html-webpack-plugin": "^5.5.0",
     "lodash": "^4.17.21",
-    "mini-css-extract-plugin": "^2.2.0",
-    "mongoose": "^5.13.5",
-    "nodemon": "2.0.12",
-    "passport": "^0.4.1",
+    "mini-css-extract-plugin": "^2.5.3",
+    "mongoose": "6.2.1",
+    "nodemon": "2.0.15",
+    "passport": "^0.5.2",
     "passport-local": "^1.0.0",
-    "postcss-loader": "^6.1.1",
+    "postcss-loader": "^6.2.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "remove-files-webpack-plugin": "^1.4.5",
-    "sass": "^1.37.5",
-    "sass-loader": "^12.1.0",
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
+    "remove-files-webpack-plugin": "^1.5.0",
+    "sass": "^1.49.7",
+    "sass-loader": "^12.4.0",
     "sass-resources-loader": "^2.2.4",
-    "style-loader": "^3.2.1",
-    "terser-webpack-plugin": "^5.1.4",
-    "webpack": "^5.49.0",
-    "webpack-bundle-analyzer": "^4.4.2",
-    "webpack-dev-middleware": "^5.0.0",
-    "webpack-hot-middleware": "^2.25.0"
+    "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.3.1",
+    "webpack": "^5.68.0",
+    "webpack-bundle-analyzer": "^4.5.0",
+    "webpack-dev-middleware": "^5.3.1",
+    "webpack-hot-middleware": "^2.25.1"
   }
 }
